move page model to model and add validation on id
