+++ g.vorobyev/lesson3/src/model/PageType.go
@@ -1,4 +1,4 @@
-package blog
+package model
 
 type PageType struct {
 	Id string `uri:"id" binding:"required"`
